+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 
 ## next
 
+- New: Alternative routes.
 - Added login statistics.
-- Fixed database migrations for H2
+- Fixed database migrations for H2.
 
 ## 0.4.1